Pull out the validate logic and reference new fx
